# backend/api/docs/swagger.yaml
basePath: /
definitions:
...
  models.UpdateUser:
    properties:
      bio:
        type: string
      imageUrl:
        type: string
      name:
        type: string
    type: object
  models.UserModel:
    properties:
      _id:
        type: string
      bio:
        type: string
      email:
        type: string
      followers:
        items:
          type: string
        type: array
      following:
        items:
          type: string
        type: array
      imageUrl:
        type: string
      name:
        type: string
      password:
        minLength: 5
        type: string
    required:
    - email
    - password
    type: object
host: localhost:5000
info:
  contact: {}
  description: 这是基于 Golang Fiber 的 REST API Swagger 文档
  title: Fiber Golang Mongo Grpc Websocket 等服务
  version: "1.0"
paths:
...
  /user/{id}/following:
    patch:
      consumes:
      - application/json
      description: 对目标用户执行关注或取消关注
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 关注/取消关注用户
      tags:
      - Users
  /user/delete:
    delete:
      consumes:
      - application/json
      description: 删除当前登录用户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - Users
  /user/getSug:
    get:
      consumes:
      - application/json
      description: 基于当前用户关注关系获取推荐用户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取推荐用户
      tags:
      - Users
  /user/getUser/{id}:
    get:
      consumes:
      - application/json
      description: 根据用户 ID 获取用户详情
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 按 ID 获取用户
      tags:
      - Users
...
  /user/update:
    patch:
      consumes:
      - application/json
      description: 更新当前登录用户资料
      parameters:
      - description: 用户更新参数
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新用户资料
      tags:
      - Users
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer 认证，格式为 "Bearer 空格 Token"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"